refactor: extract runtime uploader and captcha concerns
