+++ app/controllers/concerns/camaleon_cms/runtime_captcha_image_concern.rb
@@ -1,0 +1,38 @@
+# rubocop:disable Style/GlobalVars
+module CamaleonCms
+  module RuntimeCaptchaImageConcern
+    extend ActiveSupport::Concern
+
+    def cama_captcha_build(len = 5)
+      img = MiniMagick::Image.open(resolve_captcha_file("captcha_#{rand(12)}.jpg"))
+      text = cama_rand_str(len)
+      session[:cama_captcha] = [] if session[:cama_captcha].blank?
+      session[:cama_captcha] << text
+      img.combine_options do |c|
+        c.gravity('Center')
+        c.fill('#FFFFFF')
+        c.draw("text 0,5 #{text}")
+        c.font(resolve_captcha_file('bumpyroad.ttf'))
+        c.pointsize('30')
+      end
+    end
+
+    private
+
+    def resolve_captcha_file(filename)
+      base_dir = $camaleon_engine_dir.presence || Rails.root.to_s
+      File.join(base_dir, 'lib', 'captcha', filename)
+    end
+
+    def cama_rand_str(len = 6)
+      alphabets = [('A'..'Z').to_a].flatten!
+      alphanumerics = [('A'..'Z').to_a, ('1'..'9').to_a].flatten!
+      str = alphabets[rand(alphabets.size)]
+      (len.to_i - 1).times do
+        str << alphanumerics[rand(alphanumerics.size)]
+      end
+      str
+    end
+  end
+end
+# rubocop:enable Style/GlobalVars
